requirements.txt
Pillow==8.2.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pytesseract==0.3.7
tqdm==4.61.0
